modified file/folder structure
